+++ src/modules/decks/useCases/removeDecks/RemoveDecksUseCase.ts
@@ -2,6 +2,8 @@ import { inject, injectable } from 'tsyringe';
 
 import { IDecksRepository } from '@modules/decks/repositories/IDecksRepository';
 import IRemoveDecksDTO from "@modules/decks/dtos/IRemoveDecksDTO";
+import { AppError } from '@shared/errors/AppError';
+import { DECK_NOTFOUND, NOT_ALLOWED } from '@constants/logger';
 
 @injectable()
 export class RemoveDecksUseCase {
@@ -11,6 +13,16 @@ export class RemoveDecksUseCase {
   ) {}
 
   async execute({ deckId, userId }: IRemoveDecksDTO): Promise<void> {
+    const deck = await this.decksRepository.index({ deckId });
+
+    if (!deck) {
+      throw new AppError(DECK_NOTFOUND);
+    }
+
+    if (deck.userId !== userId) {
+      throw new AppError(NOT_ALLOWED);
+    }
+
     await this.decksRepository.remove({ deckId, userId });
   }
 }
