+++ src/modules/sessions/useCases/removeSessions/RemoveSessionsUseCase.ts
@@ -1,8 +1,9 @@
 import { inject, injectable } from 'tsyringe';
 
-import Session from '@modules/sessions/entities/Session';
 import { ISessionsRepository } from '@modules/sessions/repositories/ISessionsRepository';
 import IRemoveSessionsDTO from "@modules/sessions/dtos/IRemoveSessionsDTO";
+import { AppError } from '@shared/errors/AppError';
+import { SESSION_NOT_FOUND } from '@constants/logger';
 
 @injectable()
 export class RemoveSessionsUseCase {
@@ -12,6 +13,12 @@ export class RemoveSessionsUseCase {
   ) {}
 
   async execute({ userId, sessionId }:IRemoveSessionsDTO): Promise<void> {
-    return this.sessionsRepository.remove({ userId, sessionId });
+    const session = await this.sessionsRepository.index({ userId, sessionId });
+
+    if (!session) {
+      throw new AppError(SESSION_NOT_FOUND);
+    }
+
+    this.sessionsRepository.remove({ userId, sessionId });
   }
 }
